+++ aaem_summaries/components/residential_buildings/summary.py
@@ -14,7 +14,7 @@ from pandas import DataFrame
 
 COMPONENT_NAME = "Residential Energy Efficiency"
 DESCRIPTION = """
-    This component calculates the potential reduction heating oil by improving the efficiency of residential buildings
+    This component calculates the potential reduction in heating fuels by improving the efficiency of residential buildings
 """
 
 def generate_web_summary (web_object, community):
@@ -87,13 +87,22 @@ def generate_web_summary (web_object, community):
                                     'savings_HF')
     #~ table3[0][-1]
     year = modeled.comp_specs['data']['Year']
-    current = [{'words':'Households '+ str(int(year)) ,
+    current = [
+        {'words':'Households ('+ str(int(year)) + ')' ,
                 'value': int(modeled.comp_specs['data']['Total Occupied'])},
-        {'words':'Estimated Total Households ('+ str(int(modeled.start_year)) +')',
-        'value': modeled.init_HH},
+        
+        {'words':'Households with BEES certifcation' + \
+                ' ('+ str(int(year)) +')',
+            'value': str(modeled.comp_specs['data']['BEES Number'])},
         {'words':
-            'Estimated Households to be retofit ('+ str(int(modeled.start_year)) +')' ,
-        'value': int(modeled.opportunity_HH)},
+            'Households praticipating weatherazation of home ' + \
+            'energy rebate programs ('+ str(int(year)) +')',
+         'value': str(modeled.comp_specs['data']['Post-Retrofit Number'])},
+         
+         {'words':'Estimated Total Households (' +\
+            str(int(modeled.start_year)) +')',
+            'value': modeled.init_HH},
+        
                 ]
                 
     ## info for modeled
@@ -152,23 +161,27 @@ def create_project_details_list (project):
         A dictionary with values used by summary
     """
    
+   
+    ex_h_savings = (1 - \
+        (
+        project.proposed_HF_consumption / project.baseline_HF_consumption
+        )[0])*100
+    #~ print ex_h_savings
     return [
         {'words':'Capital cost', 
             'value': '${:,.0f}'.format(project.get_NPV_costs())},
-        {'words':'Lifetime savings', 
+        {'words':'Lifetime energy cost savings', 
             'value': '${:,.0f}'.format(project.get_NPV_benefits())},
         {'words':'Net lifetime savings', 
             'value': '${:,.0f}'.format(project.get_NPV_net_benefit())},
         {'words':'Benefit-cost ratio', 
             'value': '{:,.1f}'.format(project.get_BC_ratio())},
+        {'words':'Expected space heating savings ', 
+            'value': '{:,.2f}%'.format(ex_h_savings)},
+        {'words':
+            'Estimated households to be retofit ('\
+                + str(int(project.start_year)) +')' ,
+            'value': int(project.opportunity_HH)},
         {'words':'Estimated cost to refit household', 
             'value': '${:,.2f}/home'.format(project.refit_cost_rate)},
-        #~ {'words':'Expected Yearly Generation (kWh/year)', 
-         #~ 'value': 
-                #~ '{:,.0f}'.format(project.proposed_load *\
-                                 #~ constants.hours_per_year)},
-
-        #~ {'words':'Output per 10kW Solar PV', 
-            #~ 'value': project.comp_specs['data']\
-                                         #~ ['Output per 10kW Solar PV']},
             ]